--- backend/src/controllers/kyc/shuftipro/index.ts
@@ -1,5 +1,0 @@
-import { shuftiproPassportIssuerController } from "./zcred-passport.js";
-
-export const SHUFTI_KYC_CONTROLLER = {
-  passportIssuerWebhook: shuftiproPassportIssuerController
-};
