--- backend/src/controllers/routes/callback.route.ts
@@ -1,17 +1,0 @@
-import { Route } from "../../services/vc/vc-route.js";
-import { oauthCallbackEP } from "../../util/route.util.js";
-
-
-export const oauthCallbackRoute: Route = {
-  method: ["GET"],
-  url: oauthCallbackEP(),
-  schema: {
-    querystring: {
-      type: "object",
-      properties: {
-        code: { type: "string" },
-        state: { type: "string" },
-      },
-    },
-  },
-};
