--- backend/src/controllers/routes/persona-kyc.route.ts
@@ -1,21 +1,0 @@
-import { Route } from "../../services/vc/vc-route.js";
-
-export const personaWebhookEP = "/api/v1/kyc/callback";
-
-export const personaWebhookRoute: Route = {
-  method: ["POST"],
-  url: personaWebhookEP,
-  schema: {
-    headers: {
-      type: "object",
-      required: [
-        "Persona-Signature"
-      ],
-      properties: {
-        "Persona-Signature": {
-          type: "string"
-        }
-      }
-    }
-  }
-};
