+++ backend/src/controllers/zcred-issuer/routes.ts
@@ -1,5 +1,5 @@
 import { ZCredRoutes } from "./types.js";
-import { PASSPORT_ROUTES } from "../../issuers/zcred/passport/routes.js";
+import { PASSPORT_ROUTES } from "../../issuers/passport/routes.js";
 
 export const ZCRED_ISSUERS_ROUTES: ZCredRoutes[] = [
   PASSPORT_ROUTES,
