+++ backend/src/issuers/passport/controllers/shuftypro-kyc.controller.ts
@@ -1,4 +1,4 @@
-import { PrincipalIssuer } from "../../../issuers/zcred/index.js";
+import { PrincipalIssuer } from "../../index.js";
 import { HttpServer } from "../../../backbone/http-server.js";
 import { Injector } from "typed-inject";
 import { contextUtil } from "../../../util/context.util.js";
@@ -14,11 +14,11 @@ const TOKENS: (keyof Dependencies)[] = [
   "httpServer"
 ];
 
-const credentialType: CredentialType = "passport";
+const PASSPORT_TYPE: CredentialType = "passport";
 
-export const SHUFTI_PASSPORT_ISSUER_ENDPOINT = `/api/v1/zcred/issuers/${credentialType}/kyc/shuftipro/webhook`;
+export const SHUFTI_PASSPORT_ISSUER_ENDPOINT = `/api/v1/zcred/issuers/${PASSPORT_TYPE}/kyc/shuftipro/webhook`;
 
-export function shuftiproPassportIssuerController(injector: Injector<Dependencies>) {
+export function ShuftyproKYCPassportController(injector: Injector<Dependencies>) {
   const {
     principalIssuer,
     httpServer: { fastify }
@@ -32,7 +32,7 @@ export function shuftiproPassportIssuerController(injector: Injector<Dependencie
     },
     config: { rawBody: true },
     handler: async (req) => {
-      await principalIssuer.getIssuer(credentialType)?.handleWebhook?.(req);
+      await principalIssuer.getIssuer(PASSPORT_TYPE)?.handleWebhook?.(req);
       return { message: "ok" };
     }
   });
