--- backend/src/issuers/vc/discord-account/index.ts
@@ -1,2 +1,0 @@
-export * from "./routes.js";
-export * from "./issuer.js"
