--- backend/src/issuers/vc/github-account/index.ts
@@ -1,2 +1,0 @@
-export * from "./issuer.js"
-export * from "./routes.js"
