--- backend/src/services/vc/vc-sign-message/bitcoin-sign.service.ts
@@ -1,19 +1,0 @@
-import { SignatureService } from "./signature.service.js";
-import { verify } from "bitcoinjs-message";
-
-export async function verifyBitcoinSign(
-  signature: Uint8Array,
-  message: string,
-  address: string
-): Promise<boolean> {
-  try {
-    return verify(message, address, Buffer.from(signature));
-  } catch (e) {
-    return false;
-  }
-}
-
-export class BitcoinSignService extends SignatureService {
-  readonly didPrefix = "did:pkh:bip122:000000000019d6689c085ae165831e93";
-  protected readonly verifyFun = verifyBitcoinSign;
-}
