--- backend/src/services/vc/vc-sign-message/celo-sign.service.ts
@@ -1,5 +1,0 @@
-import { EthLikeSignService } from "./signature.service.js";
-
-export class CeloSignService extends EthLikeSignService {
-  readonly didPrefix = "did:pkh:eip155:42220";
-}
