--- backend/src/services/vc/vc-sign-message/ethereum-sign.service.ts
@@ -1,5 +1,0 @@
-import { EthLikeSignService } from "./signature.service.js";
-
-export class EthereumSignService extends EthLikeSignService {
-  readonly didPrefix = "did:pkh:eip155:1";
-}
