--- backend/src/services/vc/vc-sign-message/polygon-sign.service.ts
@@ -1,6 +1,0 @@
-import { EthLikeSignService } from "./signature.service.js";
-
-export class PolygonSignService extends EthLikeSignService {
-  readonly didPrefix = "did:pkh:eip155:137";
-
-}
