+++ backend/src/types/route.ts
@@ -1,0 +1,5 @@
+import type { RouteOptions } from "fastify/types/route.js";
+
+
+/** Base route interface each route has to implement it */
+export type Route = Omit<RouteOptions, "handler">
