--- backend/test/integration/external/vc/issue-discord-account.test.ts
@@ -1,510 +1,0 @@
-import { suite } from "uvu";
-import * as a from "uvu/assert";
-import { App } from "../../../../src/app/app.js";
-import sinon, { stub } from "sinon";
-import { canIssueEP, challengeEP, issueEP, } from "@sybil-center/sdk/util";
-import * as url from "url";
-import { configDotEnv } from "../../../../src/util/dotenv.util.js";
-import { LightMyRequestResponse } from "fastify";
-import { ethereumSupport } from "../../../support/chain/ethereum.js";
-import { solanaSupport } from "../../../support/chain/solana.js";
-import { CanIssueResp, DiscordAccountChallenge, DiscordAccountProps, DiscordAccountVC } from "@sybil-center/sdk/types";
-import { oauthCallbackEP } from "../../../../src/util/route.util.js";
-import { appSup } from "../../../support/app/index.js";
-import { bitcoinSupport } from "../../../support/chain/bitcoin.js";
-import { delay } from "../../../../src/util/delay.util.js";
-import { support } from "../../../support/index.js";
-
-type TestContext = {
-  app: App | undefined;
-  apiKey: string | undefined;
-}
-
-const test = suite<TestContext>("INTEGRATION API: issue Discord account credential test", {
-  app: undefined,
-  apiKey: undefined
-});
-
-const redirectUrl = "https://example.com/";
-const code = "code";
-const accessToken = "access_token";
-
-const discordUser = {
-  discriminator: "8822",
-  id: "123456789",
-  username: "test user name"
-};
-
-test.before(async (context) => {
-  const config = support.configPath;
-  configDotEnv({ path: config, override: true });
-  const app = await App.init();
-  const keys = await appSup.apiKeys(app);
-  context.app = app;
-  context.apiKey = keys.apiKey;
-
-  const discordAccountIssuer = app.context.resolve(
-    "discordAccountIssuer"
-  );
-  const discordService = discordAccountIssuer.discordService;
-  stub(discordService, "getAccessToken").resolves(accessToken);
-  stub(discordService, "getUser").resolves(discordUser);
-});
-
-test.after(async ({ app }) => {
-  sinon.restore();
-  await app?.close();
-});
-
-type PreIssueEntry = {
-  context: TestContext
-  custom?: Record<string, any>;
-  expirationDate?: Date;
-  subjectId: string;
-  props?: DiscordAccountProps[]
-}
-
-const preIssue = async (
-  args: PreIssueEntry
-): Promise<{
-  sessionId: string;
-  issueMessage: string
-}> => {
-  const { context: { app, apiKey } } = args;
-  const fastify = app!.context.resolve("httpServer").fastify;
-
-  const challengeResp = await fastify.inject({
-    method: "POST",
-    url: challengeEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${apiKey}`
-    },
-    payload: {
-      redirectUrl: redirectUrl,
-      custom: args.custom,
-      expirationDate: args.expirationDate,
-      subjectId: args.subjectId,
-      props: args.props
-    }
-  });
-  a.is(challengeResp.statusCode, 200,
-    "payload response status code is not 200");
-  const { sessionId, issueMessage, authUrl } =
-    JSON.parse(challengeResp.body) as DiscordAccountChallenge;
-
-  const { query } = url.parse(authUrl, true);
-  const state = query.state as string;
-
-  const canIssueBeforeResp = await fastify.inject({
-    method: "GET",
-    url: canIssueEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${apiKey}`
-    },
-    query: {
-      sessionId: sessionId
-    }
-  });
-  a.is(
-    canIssueBeforeResp.statusCode,
-    200,
-    "can issue before callback resp status is not 200"
-  );
-  const { canIssue: canIssueBefore } =
-    JSON.parse(canIssueBeforeResp.body) as CanIssueResp;
-  a.is(canIssueBefore, false, "can issue before callback is not false");
-
-  const callbackResp = await fastify.inject({
-    method: "GET",
-    url: oauthCallbackEP(),
-    query: {
-      code: code,
-      state: state
-    }
-  });
-  a.is(
-    callbackResp.statusCode,
-    302,
-    "callback response status code is not 302"
-  );
-  const redirectedTo = callbackResp.headers.location as string;
-  a.is(redirectedTo, redirectUrl, "redirect url is not matched");
-
-  const canIssueAfterResp = await fastify.inject({
-    method: "GET",
-    url: canIssueEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${apiKey}`
-    },
-    query: {
-      sessionId: sessionId
-    }
-  });
-  a.is(
-    canIssueAfterResp.statusCode, 200,
-    "can issue after callback resp status is not 200"
-  );
-  const { canIssue: canIssueAfter } =
-    JSON.parse(canIssueAfterResp.body) as CanIssueResp;
-  a.is(canIssueAfter, true, "can issue after callback is not true");
-
-  return {
-    sessionId: sessionId,
-    issueMessage: issueMessage
-  };
-};
-
-const assertIssueResp = async (
-  issueResp: LightMyRequestResponse,
-  subjectDID: string,
-  context: TestContext,
-) => {
-  const issuerId = context.app!.context.resolve("didService").id;
-  a.is(issueResp.statusCode, 200,
-    `issue resp status code is not 200. error: ${issueResp.body}`);
-
-  const vc = JSON.parse(issueResp.body) as DiscordAccountVC;
-  const {
-    discriminator,
-    id: discordId,
-    username
-  } = vc.credentialSubject.discord;
-
-  a.is(vc.issuer.id, issuerId, "issuer id is not matched");
-  a.is(vc.type.includes("DiscordAccount"), true);
-  a.is(
-    vc.credentialSubject.id, subjectDID,
-    "vc credential subject id is not matched"
-  );
-  a.is(
-    discriminator,
-    discordUser.discriminator,
-    "discord user discriminator is not matched"
-  );
-  a.is(discordId, discordUser.id, "discord user id is not matched");
-  a.is(username, discordUser.username, "discord username is not matched");
-};
-
-
-const assertSessionDeleted = (sessionId: string, context: TestContext) => {
-  // @ts-expect-error
-  const { sessionCache } = context.app!.context.resolve("discordAccountIssuer");
-  a.throws(() => {
-    sessionCache.get(sessionId);
-  }, "session is not deleted");
-};
-
-test("should issue discord ownership credential with ethereum did-pkh", async (context) => {
-  const ethDidPkh = ethereumSupport.info.ethereum.didPkh;
-  const fastify = context.app!.context.resolve("httpServer").fastify;
-  const { issueMessage, sessionId } = await preIssue({
-    subjectId: ethDidPkh,
-    context: context
-  });
-  const signature = await ethereumSupport.sign(issueMessage);
-
-  const vcResponse = await fastify.inject({
-    method: "POST",
-    url: issueEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${context.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: signature,
-    }
-  });
-  await assertIssueResp(vcResponse, ethDidPkh, context);
-  assertSessionDeleted(sessionId, context);
-  const credential = JSON.parse(vcResponse.body);
-  await appSup.verifyCredential({ credential: credential, app: context.app! });
-});
-
-test("should issue discord ownership credential with solana did-pkh", async (context) => {
-  const {
-    didPkh: solanaDidPkh,
-  } = solanaSupport.info;
-  const { fastify } = context.app!.context.resolve("httpServer");
-
-  const { sessionId, issueMessage } = await preIssue({ subjectId: solanaDidPkh, context });
-  const signature = await solanaSupport.sign(issueMessage);
-
-  const vcResponse = await fastify.inject({
-    method: "POST",
-    url: issueEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${context.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: signature,
-    }
-  });
-  await assertIssueResp(vcResponse, solanaDidPkh, context);
-  assertSessionDeleted(sessionId, context);
-  const credential = JSON.parse(vcResponse.body);
-  await appSup.verifyCredential({ credential: credential, app: context.app! });
-});
-
-test("should issue discord ownership credential with bitcoin did-pkh", async (context) => {
-  const { didPkh: bitcoinDidPkh, } = bitcoinSupport.info;
-  const { fastify } = context.app!.context.resolve("httpServer");
-
-  const { sessionId, issueMessage } = await preIssue({ subjectId: bitcoinDidPkh, context });
-  const signature = await bitcoinSupport.sing(issueMessage);
-
-  const vcResponse = await fastify.inject({
-    method: "POST",
-    url: issueEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${context.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: signature
-    }
-  });
-  await assertIssueResp(vcResponse, bitcoinDidPkh, context);
-  assertSessionDeleted(sessionId, context);
-  const credential = JSON.parse(vcResponse.body);
-  await appSup.verifyCredential({ credential: credential, app: context.app! });
-});
-
-test("should redirect to default page after authorization", async (context) => {
-  const { didPkh } = ethereumSupport.info.ethereum;
-  const { fastify } = context.app!.context.resolve("httpServer");
-  const config = context.app!.context.resolve("config");
-  const challengeResp = await fastify.inject({
-    method: "POST",
-    url: challengeEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${context.apiKey}`
-    },
-    payload: {
-      subjectId: didPkh
-    }
-  });
-  a.is(challengeResp.statusCode, 200, "payload resp status code is not 200");
-  const { authUrl } = JSON.parse(
-    challengeResp.body
-  ) as DiscordAccountChallenge;
-  const { query } = url.parse(authUrl, true);
-  const state = query.state as string;
-
-  const callbackResp = await fastify.inject({
-    method: "GET",
-    url: oauthCallbackEP(),
-    query: {
-      code: code,
-      state: state
-    }
-  });
-  a.is(
-    callbackResp.statusCode,
-    302,
-    "callback response status code is not 302"
-  );
-  const defaultRedirectUrl = callbackResp.headers.location as string;
-  a.ok(
-    defaultRedirectUrl,
-    "default redirect url after authorization is not present"
-  );
-  a.is(
-    defaultRedirectUrl,
-    new URL("/oauth/authorized", config.pathToExposeDomain).href,
-    "default redirect url after authorization is not matched"
-  );
-});
-
-test("should issue vc with custom properties", async (context) => {
-  const custom = { hello: { test: "test", world: "world" } };
-  const { didPkh } = ethereumSupport.info.ethereum;
-  const { fastify } = context.app!.context.resolve("httpServer");
-  const { sessionId, issueMessage } = await preIssue({
-    subjectId: didPkh,
-    custom: custom,
-    context
-  });
-  const signature = await ethereumSupport.sign(issueMessage);
-  const issueResp = await fastify.inject({
-    method: "POST",
-    url: issueEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${context.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: signature,
-    }
-  });
-  a.is(issueResp.statusCode, 200, `issue resp status code is not 200. error: ${issueResp.body}`);
-  await assertIssueResp(issueResp, didPkh, context);
-  const vc = JSON.parse(issueResp.body) as DiscordAccountVC;
-  const vcCustom = vc.credentialSubject.custom;
-  a.ok(vcCustom, "custom property is not present");
-  a.ok(vcCustom.hello, "custom hello is not present");
-  a.is(vcCustom.hello.test, "test", "custom property is not matched");
-  a.is(vcCustom.hello.world, "world", "custom property is not matched");
-  assertSessionDeleted(sessionId, context);
-  await appSup.verifyCredential({ credential: vc, app: context.app! });
-});
-
-test("should not find Discord code", async (context) => {
-  const { didPkh } = ethereumSupport.info.ethereum;
-  const fastify = context.app!.context.resolve("httpServer").fastify;
-  await fastify.ready();
-  const challengeResp = await fastify.inject({
-    method: "POST",
-    url: challengeEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${context.apiKey}`
-    },
-    payload: {
-      redirectUrl: redirectUrl,
-      subjectId: didPkh
-    }
-  });
-  a.is(
-    challengeResp.statusCode, 200,
-    "challenge response status code is not 200"
-  );
-
-  const { sessionId, issueMessage } = JSON.parse(
-    challengeResp.body
-  ) as DiscordAccountChallenge;
-  const signature = await ethereumSupport.sign(issueMessage);
-
-  const errResp = await fastify.inject({
-    method: "POST",
-    url: issueEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${context.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: signature,
-    }
-  });
-  a.is(errResp.statusCode, 400, "error response status code is not 400");
-  const { message } = JSON.parse(errResp.body);
-  a.is(message, "Discord processing your authorization. Wait!",
-    "error message not matched");
-  await fastify.ready();
-});
-
-test("issue discord account credential with expiration date", async (context) => {
-  const { didPkh: subjectDID } = ethereumSupport.info.ethereum;
-  const { fastify } = context.app!.context.resolve("httpServer");
-  const expirationDate = new Date();
-  const { sessionId, issueMessage } = await preIssue({
-    subjectId: subjectDID,
-    expirationDate: expirationDate,
-    context
-  });
-  const signature = await ethereumSupport.sign(issueMessage);
-
-  const issueResp = await fastify.inject({
-    method: "POST",
-    url: issueEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${context.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: signature,
-    }
-  });
-  await assertIssueResp(issueResp, subjectDID, context);
-  assertSessionDeleted(sessionId, context);
-  const credential = JSON.parse(issueResp.body);
-  a.ok(credential.expirationDate, "credential expiration date is not present");
-  a.is(
-    credential.expirationDate,
-    expirationDate.toISOString(),
-    "credential expiration date is not matched");
-
-  await delay(20);
-  await appSup.verifyCredential({
-    credential: credential,
-    app: context.app!,
-    shouldVerified: false
-  });
-});
-
-test("not valid date-time format for expiration date", async (context) => {
-  const { fastify } = context.app!.context.resolve("httpServer");
-  const ethDidPkh = ethereumSupport.info.ethereum.didPkh;
-  const errResp = await fastify.inject({
-    method: "POST",
-    url: challengeEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${context.apiKey}`
-    },
-    payload: {
-      redirectUrl: redirectUrl,
-      expirationDate: "not a date",
-      subjectId: ethDidPkh
-    }
-  });
-  a.is(errResp.statusCode, 400,
-    `invalid expiration date is ok. error: ${errResp.body}`);
-});
-
-test("should issue eth account credential without props", async (context) => {
-  const fastify = context.app!.context.resolve("httpServer").fastify;
-  const { didPkh } = ethereumSupport.info.ethereum;
-  const {
-    sessionId,
-    issueMessage
-  } = await preIssue({ subjectId: didPkh, props: [], context });
-  const signature = await ethereumSupport.sign(issueMessage);
-  const issueResp = await fastify.inject({
-    method: "POST",
-    url: issueEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${context.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: signature,
-    }
-  });
-  a.is(issueResp.statusCode, 200, `issue resp fail. error: ${issueResp.body}`);
-  const credential = JSON.parse(issueResp.body) as DiscordAccountVC;
-  a.ok(credential.credentialSubject.discord, `subject discord field is empty`);
-  a.not.ok(credential.credentialSubject.discord.id);
-  a.not.ok(credential.credentialSubject.discord.username);
-  a.not.ok(credential.credentialSubject.discord.discriminator);
-});
-
-test("should issue eth account credential with only one prop", async (ctx) => {
-  const fastify = ctx.app!.context.resolve("httpServer").fastify;
-  const { didPkh } = ethereumSupport.info.ethereum;
-  const {
-    sessionId,
-    issueMessage
-  } = await preIssue({ subjectId: didPkh, props: ["username"], context: ctx });
-  const signature = await ethereumSupport.sign(issueMessage);
-  const issueResp = await fastify.inject({
-    method: "POST",
-    url: issueEP("DiscordAccount"),
-    headers: {
-      Authorization: `Bearer ${ctx.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: signature,
-    }
-  });
-  a.is(issueResp.statusCode, 200, `issue resp fail. error: ${issueResp.body}`);
-  const credential = JSON.parse(issueResp.body) as DiscordAccountVC;
-  a.ok(credential.credentialSubject.discord, `subject discord field is empty`);
-  a.is(
-    credential.credentialSubject.discord.username, discordUser.username,
-    `subject discord user nate is not matched`
-  );
-  a.not.ok(credential.credentialSubject.discord.id, `discord field contains id prop`);
-  a.not.ok(credential.credentialSubject.discord.discriminator, `discord field contains discriminator`);
-});
-
-// test.run();
