--- backend/test/integration/external/vc/issue-eth-account.test.ts
@@ -1,313 +1,0 @@
-import { suite } from "uvu";
-import * as a from "uvu/assert";
-import { App } from "../../../../src/app/app.js";
-import sinon from "sinon";
-import { challengeEP, issueEP } from "@sybil-center/sdk/util";
-import { configDotEnv } from "../../../../src/util/dotenv.util.js";
-import { ethereumSupport } from "../../../support/chain/ethereum.js";
-import { EthAccountChallenge, EthAccountProps, EthAccountVC } from "@sybil-center/sdk/types";
-import { LightMyRequestResponse } from "fastify";
-import { appSup } from "../../../support/app/index.js";
-import { delay } from "../../../../src/util/delay.util.js";
-import { support } from "../../../support/index.js";
-
-type TestContext = {
-  app: App | undefined;
-  apiKey: string | undefined;
-}
-
-const test = suite<TestContext>("INTEGRATION API: issue Ethereum account credential test", {
-  app: undefined,
-  apiKey: undefined
-});
-
-test.before(async (ctx) => {
-  configDotEnv({ path: support.configPath, override: true });
-  const app = await App.init();
-  const keys = await appSup.apiKeys(app);
-  ctx.apiKey = keys.apiKey;
-  ctx.app = app;
-});
-
-test.after(async (ctx) => {
-  sinon.restore();
-  await ctx.app?.close();
-});
-
-type PreIssueEntry = {
-  subjectId: string;
-  custom?: Record<string, any>,
-  expirationDate?: Date,
-  props?: EthAccountProps[],
-  ctx: TestContext;
-}
-const preIssue = async (
-  args: PreIssueEntry
-): Promise<{
-  sessionId: string;
-  issueMessage: string;
-}> => {
-  const { app, apiKey } = args.ctx;
-  const fastify = app!.context.resolve("httpServer").fastify;
-  const challengeResp = await fastify.inject({
-    method: "POST",
-    url: challengeEP("EthereumAccount"),
-    headers: {
-      Authorization: `Bearer ${apiKey}`
-    },
-    payload: {
-      subjectId: args.subjectId,
-      custom: args.custom,
-      expirationDate: args.expirationDate,
-      props: args.props
-    }
-  });
-  a.is(
-    challengeResp.statusCode, 200,
-    "challenge response status code is not 200"
-  );
-  const { sessionId, issueMessage } = JSON.parse(
-    challengeResp.body
-  ) as EthAccountChallenge;
-  return { sessionId, issueMessage: issueMessage };
-};
-
-type AssertIssueRespArgs = {
-  issueResp: LightMyRequestResponse,
-  subjectDID: string,
-  ethereumAddress: string;
-  ctx: TestContext;
-}
-
-const assertIssueResp = async ({
-  issueResp,
-  subjectDID,
-  ethereumAddress,
-  ctx: { app }
-}: AssertIssueRespArgs) => {
-  const issuerDID = app!.context.resolve("didService").id;
-  a.is(
-    issueResp.statusCode, 200,
-    `issue resp status code is not 200. error: ${issueResp.body}`);
-  const credential = JSON.parse(issueResp.body) as EthAccountVC;
-  a.is(
-    credential.issuer.id, issuerDID,
-    "issuer id is not matched"
-  );
-  a.is(
-    credential.credentialSubject.id, subjectDID,
-    "credential subject did is not matched"
-  );
-  a.is(
-    credential.type[0], "VerifiableCredential",
-    "credential type first item is not correct"
-  );
-  a.is(
-    credential.type[1], "EthereumAccount",
-    "credential type second item is not correct"
-  );
-  a.is(
-    credential.credentialSubject.ethereum.address, ethereumAddress,
-    "credential subject ethereum address is not matched"
-  );
-  a.is(
-    credential.credentialSubject.ethereum.chainId, "eip155:1",
-    "credential subject ethereum sign id is not matched"
-  );
-};
-
-const assertSessionDeleted = (sessionId: string, ctx: TestContext) => {
-  // @ts-expect-error
-  let { sessionCache } = ctx.app!.context.resolve("ethereumAccountIssuer");
-  a.throws(() => {
-    sessionCache.get(sessionId);
-  }, "session is not deleted");
-};
-
-test("should issue ethereum account vc", async (ctx) => {
-  const { didPkh: subjectDID, address } = ethereumSupport.info.ethereum;
-  const fastify = ctx.app!.context.resolve("httpServer").fastify;
-  const { issueMessage, sessionId } = await preIssue({ subjectId: subjectDID, ctx });
-  const signature = await ethereumSupport.sign(issueMessage);
-  const issueResp = await fastify.inject({
-    method: "POST",
-    url: issueEP("EthereumAccount"),
-    headers: {
-      Authorization: `Bearer ${ctx.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: signature,
-    }
-  });
-  await assertIssueResp({ issueResp, subjectDID, ethereumAddress: address, ctx });
-  assertSessionDeleted(sessionId, ctx);
-  const credential = JSON.parse(issueResp.body);
-  await appSup.verifyCredential({ credential: credential, app: ctx.app! });
-});
-
-test("should not issue credential because not valid signature", async (ctx) => {
-  const { didPkh } = ethereumSupport.info.ethereum;
-  const fastify = ctx.app!.context.resolve("httpServer").fastify;
-  const challengeResp = await fastify.inject({
-    method: "POST",
-    url: challengeEP("EthereumAccount"),
-    headers: {
-      Authorization: `Bearer ${ctx.apiKey}`
-    },
-    payload: {
-      subjectId: didPkh
-    }
-  });
-  a.is(
-    challengeResp.statusCode, 200,
-    `payload response status code is not 200. error: ${challengeResp.body}`
-  );
-  const {
-    sessionId,
-    issueMessage
-  } = JSON.parse(challengeResp.body) as EthAccountChallenge;
-  const errResp = await fastify.inject({
-    method: "POST",
-    url: issueEP("EthereumAccount"),
-    headers: {
-      Authorization: `Bearer ${ctx.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: issueMessage,
-    }
-  });
-  a.is(
-    errResp.statusCode, 400,
-    "error response status code is not 400"
-  );
-});
-
-test("should issue ethereum account credential with custom property", async (ctx) => {
-  const {
-    didPkh: subjectDID,
-    address: ethAddress,
-  } = ethereumSupport.info.ethereum;
-  const custom = { test: { hello: "world" } };
-  const { fastify } = ctx.app!.context.resolve("httpServer");
-
-  const { issueMessage, sessionId } = await preIssue({
-    subjectId: subjectDID,
-    custom: custom,
-    ctx
-  });
-  const signature = await ethereumSupport.sign(issueMessage);
-  const issueResp = await fastify.inject({
-    method: "POST",
-    url: issueEP("EthereumAccount"),
-    headers: {
-      Authorization: `Bearer ${ctx.apiKey}`
-    },
-    payload: {
-      signature: signature,
-      sessionId: sessionId,
-    }
-  });
-  await assertIssueResp({ issueResp, subjectDID, ethereumAddress: ethAddress, ctx });
-  const credential = JSON.parse(issueResp.body) as EthAccountVC;
-  const { custom: vcCustom } = credential.credentialSubject;
-  a.ok(vcCustom, "custom is not present");
-  a.ok(vcCustom["test"], "custom.test is not present");
-  a.ok(vcCustom["test"]?.hello, "custom.test.hello is not present");
-  a.is(
-    vcCustom["test"]?.hello, "world",
-    "hello custom property is not matched"
-  );
-  assertSessionDeleted(sessionId, ctx);
-  await appSup.verifyCredential({ credential: credential, app: ctx.app! });
-});
-
-test("issue eth account credential with expiration date", async (ctx) => {
-  const fastify = ctx.app!.context.resolve("httpServer").fastify;
-  const {
-    didPkh: subjectDID,
-    address: ethereumAddress,
-  } = ethereumSupport.info.ethereum;
-  const expirationDate = new Date();
-  const { sessionId, issueMessage } = await preIssue({
-    subjectId: subjectDID,
-    expirationDate: expirationDate,
-    ctx
-  });
-  const signature = await ethereumSupport.sign(issueMessage);
-  const issueResp = await fastify.inject({
-    method: "POST",
-    url: issueEP("EthereumAccount"),
-    headers: {
-      Authorization: `Bearer ${ctx.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: signature
-    }
-  });
-  await assertIssueResp({ issueResp, ethereumAddress, subjectDID, ctx });
-  const credential = JSON.parse(issueResp.body) as EthAccountVC;
-  a.is(
-    credential.expirationDate, expirationDate.toISOString(),
-    "credential expiration date is not matched"
-  );
-  assertSessionDeleted(sessionId, ctx);
-  await delay(20);
-  await appSup.verifyCredential({
-    credential: credential,
-    app: ctx.app!,
-    shouldVerified: false
-  });
-});
-
-test("should issue eth credential without props", async (ctx) => {
-  const fastify = ctx.app!.context.resolve("httpServer").fastify;
-  const { didPkh } = ethereumSupport.info.ethereum;
-  const { issueMessage, sessionId } = await preIssue({ subjectId: didPkh, props: [], ctx });
-  const signature = await ethereumSupport.sign(issueMessage);
-  const issueResp = await fastify.inject({
-    method: "POST",
-    path: issueEP("EthereumAccount"),
-    headers: {
-      Authorization: `Bearer ${ctx.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: signature,
-    }
-  });
-  a.is(issueResp.statusCode, 200, `issue resp fail. error ${issueResp.body}`);
-  const credential = JSON.parse(issueResp.body) as EthAccountVC;
-  const ethereum = credential.credentialSubject.ethereum;
-  a.ok(ethereum, `subject ethereum field is undefined`);
-  a.not.ok(ethereum.chainId, `subject ethereum field has to be empty`);
-  a.not.ok(ethereum.address, `subject ethereum field has to be empty`);
-});
-
-test("should issue eth account credential with only one prop", async (ctx) => {
-  const fastify = ctx.app!.context.resolve("httpServer").fastify;
-  const { didPkh, address } = ethereumSupport.info.ethereum;
-  const { sessionId, issueMessage } = await preIssue({ subjectId: didPkh, props: ["address"], ctx });
-  const signature = await ethereumSupport.sign(issueMessage);
-  const issueResp = await fastify.inject({
-    method: "POST",
-    path: issueEP("EthereumAccount"),
-    headers: {
-      Authorization: `Bearer ${ctx.apiKey}`
-    },
-    payload: {
-      sessionId: sessionId,
-      signature: signature,
-    }
-  });
-  a.is(issueResp.statusCode, 200, `issue reps fail. error: ${issueResp.body}`);
-  const credential = JSON.parse(issueResp.body) as EthAccountVC;
-  const ethereum = credential.credentialSubject.ethereum;
-  a.ok(ethereum, `subject ethereum field is empty`);
-  a.is(ethereum.address, address, `ethereum address is not matched`);
-  a.not.ok(ethereum.chainId, `subject ethereum.chainId must be empty`);
-});
-
-// test.run();
