+++ backend/test/integration/external/zcred/passport-issuer.test.ts
@@ -1,10 +1,10 @@
 import { suite } from "uvu";
-import { App } from "../../../../../src/app/app.js";
-import { configDotEnv } from "../../../../../src/util/dotenv.util.js";
-import { support } from "../../../../support/index.js";
+import { App } from "../../../../src/app/app.js";
+import { configDotEnv } from "../../../../src/util/dotenv.util.js";
+import { support } from "../../../support/index.js";
 import { FastifyInstance } from "fastify";
-import { Config } from "../../../../../src/backbone/config.js";
-import { PassportIssuer } from "../../../../../src/issuers/zcred/passport/index.js";
+import { Config } from "../../../../src/backbone/config.js";
+import { PassportIssuer } from "../../../../src/issuers/passport/index.js";
 import { Challenge, ChallengeReq, HttpCredential, Identifier, Info, IssueReq, StrictId, zcredjs, } from "@zcredjs/core";
 import { EthSignature } from "@zcredjs/ethereum";
 import * as o1js from "o1js";
@@ -12,17 +12,17 @@ import { Field, Scalar, Signature } from "o1js";
 import { MinaPoseidonPastaVerifier } from "@zcredjs/mina";
 import * as a from "uvu/assert";
 import sinon from "sinon";
-import { ICredentialSignProver } from "../../../../../src/services/credential-provers/type.js";
-import { ethereumSupport } from "../../../../support/chain/ethereum.js";
+import { ICredentialSignProver } from "../../../../src/services/credential-provers/type.js";
+import { ethereumSupport } from "../../../support/chain/ethereum.js";
 import { ethers } from "ethers";
-import { minaSupport } from "../../../../support/chain/mina.js";
+import { minaSupport } from "../../../support/chain/mina.js";
 import Client from "mina-signer";
-import { DIDService } from "../../../../../src/services/did.service.js";
+import { DIDService } from "../../../../src/services/did.service.js";
 import sortKeys from "sort-keys";
-import { toJWTPayload } from "../../../../../src/util/jwt.util.js";
-import { sybil } from "../../../../../src/services/sybiljs/index.js";
-import { IPassportKYCService } from "../../../../../src/issuers/zcred/passport/types.js";
-import { PassportCredential } from "../../../../../src/services/sybiljs/passport/types.js";
+import { toJWTPayload } from "../../../../src/util/jwt.util.js";
+import { sybil } from "../../../../src/services/sybiljs/index.js";
+import { IPassportKYCService } from "../../../../src/issuers/passport/types.js";
+import { PassportCredential } from "../../../../src/services/sybiljs/passport/types.js";
 
 const test = suite("INTEGRATION API: ZCred passport issuer");
