--- backend/test/unit/services/vc/sign-message/bitcoin-sign.service.test.ts
@@ -1,23 +1,0 @@
-import { suite } from "uvu";
-import * as assert from "uvu/assert";
-import { BitcoinSignService } from "../../../../../src/services/vc/vc-sign-message/bitcoin-sign.service.js";
-import { bitcoinSupport } from "../../../../support/chain/bitcoin.js";
-
-const test = suite("UNIT: bitcoin sign test");
-
-const bitcoinChain = new BitcoinSignService();
-const { didPkhPrefix, address } = bitcoinSupport.info;
-
-const message = "test";
-const signature = await bitcoinSupport.sing(message);
-
-test("should correct verify signature", async () => {
-  await bitcoinChain.verify({ signature, message, address });
-});
-
-test("should correct get did-pkh", async () => {
-  const didPkh = await bitcoinChain.did({ signature, message, address });
-  assert.is(didPkh, `${didPkhPrefix}:${address}`, "did-pkh is not matched");
-});
-
-test.run();
