--- backend/test/unit/services/vc/sign-message/ethereum-sign.service.test.ts
@@ -1,22 +1,0 @@
-import { suite } from "uvu";
-import * as assert from "uvu/assert";
-import { EthereumSignService } from "../../../../../src/services/vc/vc-sign-message/ethereum-sign.service.js";
-import { ethereumSupport } from "../../../../support/chain/ethereum.js";
-
-const test = suite("UNIT: ethereum sign test");
-
-const ethereumChain = new EthereumSignService();
-const { address, didPkhPrefix } = ethereumSupport.info.ethereum;
-const message = "test";
-const signature = await ethereumSupport.sign(message);
-
-test("should verify signature", async () => {
-  await ethereumChain.verify({ signature, message, address });
-});
-
-test("should correct get did-pkh", async () => {
-  const didPkh = await ethereumChain.did({ signature, message, address });
-  assert.is(didPkh, `${didPkhPrefix}:${address}`, "did-pkh is not matched");
-});
-
-test.run();
