--- backend/test/unit/services/vc/sign-message/polygon-signservice.test.ts
@@ -1,23 +1,0 @@
-import { suite } from "uvu";
-import * as assert from "uvu/assert";
-import { PolygonSignService } from "../../../../../src/services/vc/vc-sign-message/polygon-sign.service.js";
-import { ethereumSupport } from "../../../../support/chain/ethereum.js";
-
-const test = suite("UNIT: polygon sign test");
-
-const polygonChain = new PolygonSignService();
-
-const { address, didPkhPrefix } = ethereumSupport.info.polygon;
-const message = "test";
-const signature = await ethereumSupport.sign(message);
-
-test("should correct verify message", async () => {
-  await polygonChain.verify({ signature, message, address });
-});
-
-test("should correct get did-pkh", async () => {
-  const didPkh = await polygonChain.did({ signature, message, address });
-  assert.is(didPkh, `${didPkhPrefix}:${address}`);
-});
-
-test.run();
